ガイドボタンのスクリプトの修正
