+++ Big Wave prototype/Assets/Script/UIScript/ButtonIcon/GuideButtonIconFullScreen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eaf03e4b0033ad94eb2f14470364071f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
