+++ Big Wave prototype/Assets/Script/UIScript/ButtonIcon/GuideButtonIconFullScreen/ButtonDisplays.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 46bf5513184f6a541b95c4a4cd5e1576
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
