+++ Big Wave prototype/Assets/Script/UIScript/ButtonIcon/GuideButtonIconFullScreen/GuideButtonIconFullScreen.cs.meta
